refact: 添加配置管理功能，重构数据库连接逻辑，简化错误处理，增强代码一致性和可读性。同时新增 clippy 和 rustfmt 配置文件以提升代码质量。

+++ src-tauri/src/types/scheduler.rs
@@ -26,7 +26,7 @@ pub struct ScheduledJobUpdate {
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct ScheduledJobResponse {
-    pub id: Option<u32>,
+    pub id: Option<i64>,
     pub job_id: String,
     pub name: String,
     pub cron_expression: String,
